backend/requirements.txt

fastapi
uvicorn
openai
firebase-admin
python-dotenv
pydantic
